Extract currency convert form to a new component
