+++ pages/exchange.tsx
@@ -1,64 +1,10 @@
 import Head from "next/head"
-import React, { useState } from "react"
-import axios from "axios"
-import Swal from "sweetalert2"
-import Currency from "../components/currency"
+import React from "react"
 import Emoji from "../components/emoji"
-import { RequestData } from "./api/exchange"
+import CurrencyConvertForm from "../components/exchange/currencyConvertForm"
 import styles from "../styles/pages/Exchange.module.css"
 
-const ErrorToast = Swal.mixin({
-  icon: "error",
-  toast: true,
-  position: "center",
-  showConfirmButton: false,
-  timer: 3000,
-  timerProgressBar: true,
-  width: "42em",
-})
-
 export default function Exchange() {
-  const [amount, setAmount] = useState<Currency>(new Currency(0))
-  const [fromCurrency, setFromCurrency] = useState<string>("")
-  const [toCurrency, setToCurrency] = useState<string>("")
-
-  async function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault()
-    if (fromCurrency === toCurrency) {
-      await ErrorToast.fire({
-        title: "Invalid currency chosen",
-        text: "Currency chosen for conversion should be different",
-      })
-      return
-    }
-    const body: RequestData = {
-      fromCurrencyCode: fromCurrency,
-      toCurrencyCode: toCurrency,
-      amount: amount.getAmountInDollars(),
-    }
-    const response = await axios.post("/api/exchange", body)
-    console.log(JSON.stringify(response.data))
-    // TODO Display results
-  }
-
-  function handleFromCurrencyChange(event: React.ChangeEvent<HTMLSelectElement>) {
-    setFromCurrency(event.target.value)
-  }
-
-  function handleToCurrencyChange(event: React.ChangeEvent<HTMLSelectElement>) {
-    setToCurrency(event.target.value)
-  }
-
-  function handleChangeAmount(event: React.ChangeEvent<HTMLInputElement>): void {
-    const input: string = event.target.value
-    if (input.trim() === "") {
-      setAmount(new Currency(0))
-      return
-    }
-    const amountInCents: number = parseFloat(input) * 100
-    setAmount(new Currency(amountInCents))
-  }
-
   return (
     <div className={styles.container}>
       <Head>
@@ -71,59 +17,7 @@ export default function Exchange() {
         <Emoji symbol="🌎" label="world" />
       </h1>
 
-      <form action="/api/exchange"
-            method="post"
-            onSubmit={handleSubmit}
-            className={styles.currencyForm}
-            aria-label="exchange-currency-form"
-      >
-        <div className={styles.currencyFormInput}>
-          <label htmlFor="conversion-amount">Conversion Amount</label>
-          <input type="number"
-                 required
-                 id="conversion-amount"
-                 min="0.01"
-                 max="99999999"
-                 step="0.01"
-                 value={String(amount.getAmountInDollars())}
-                 onChange={handleChangeAmount}
-                 className={styles.input}
-          />
-        </div>
-        <div className={styles.currencyFormInput}>
-          <label htmlFor="from-currency-select">From</label>
-          <select required
-                  value={fromCurrency}
-                  onChange={handleFromCurrencyChange}
-                  aria-label="from-currency-select"
-                  className={styles.input}
-          >
-            <option value="" aria-label="default-from-currency" disabled hidden>
-              -- select an option --
-            </option>
-            <option value="SGD" aria-label="from-SGD">SGD - Singapore Dollar</option>
-            <option value="USD" aria-label="from-USD">USD - United States Dollar</option>
-          </select>
-        </div>
-        <div className={styles.currencyFormInput}>
-          <label htmlFor="to-currency-select">To</label>
-          <select required
-                  value={toCurrency}
-                  onChange={handleToCurrencyChange}
-                  aria-label="to-currency-select"
-                  className={styles.input}
-          >
-            <option value="" aria-label="default-to-currency" disabled hidden>
-              -- select an option --
-            </option>
-            <option value="SGD" aria-label="to-SGD">SGD - Singapore Dollar</option>
-            <option value="USD" aria-label="to-USD">USD - United States Dollar</option>
-          </select>
-        </div>
-        <button type="submit" className={styles.currencyFormButton}>
-          Convert
-        </button>
-      </form>
+      <CurrencyConvertForm />
     </div>
   )
 }
