+++ styles/components/exchange/CurrencyConvertForm.module.css
@@ -1,0 +1,32 @@
+.currencyForm {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: 1rem;
+  width: 80%;
+}
+
+.currencyFormInput {
+  display: flex;
+  flex-direction: column;
+}
+
+.currencyFormButton {
+  padding: 0 2em;
+}
+
+.input {
+  font-size: 1rem;
+  padding: 0.2em 1em;
+}
+
+@media (max-width: 600px) {
+  .currencyForm {
+    flex-direction: column;
+    width: 90%;
+  }
+
+  .currencyFormButton {
+    padding: 0.8em 2em;
+  }
+}
